+++ .gitignore
@@ -3,4 +3,3 @@ build/
 node_modules
 
 package-lock.json
-tester
