+++ README.md
@@ -12,3 +12,21 @@ Start in dev mode:
 Start in prod mode:
 - `npm run build` or `yarn build`
 - `npm start` or `yarn start`
+
+# Run with tester (https://github.com/introduction-to-gamedev/reversi-ai)
+
+### Linux
+
+Build project:
+- `npm run build` or `yarn build`
+
+pass `runTester.sh` as --command argument value when start tester CLI.
+
+### Windows
+
+Build project:
+- `npm run build` or `yarn build`
+
+Create runnable cmd/exe file that will execute command `node build\index.js`
+
+Pass created runnable file name as --command argument value when start tester CLI.
