+++ package.json
@@ -2,17 +2,14 @@
   "name": "reversi-game",
   "version": "1.0.0",
   "description": "",
-  "main": "src/indexOld.ts",
+  "main": "src/index.ts",
   "scripts": {
     "start": "node build/index.js",
     "start:dev": "ts-node src/index.ts",
     "build": "tsc",
     "lint": "eslint src/**/*.ts",
     "lint:fix": "npm run lint -- --fix",
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "bot": "dotnet tester/reversi-ai-tester.dll -c --command \"./tester/runTest.sh\"",
-    "bot:prod": "dotnet tester/reversi-ai-tester.dll -c --command \"./tester/runTestProd.sh\"",
-    "bot:single": "dotnet tester/reversi-ai-tester.dll -c -s --command \"./tester/runTest.sh\""
+    "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
     "type": "git",
