+++ runTester.sh
@@ -1,0 +1,2 @@
+#!/usr/bin/env bash
+node ./build/index.js
