+++ src/gameInfo.interface.ts
@@ -1,0 +1,7 @@
+import {Color} from './model/color.enum';
+
+export interface GameInfo {
+  color: Color,
+  blackHole: string,
+  firstOpponentMove: string
+}
