+++ src/model/color.enum.ts
@@ -1,0 +1,4 @@
+export enum Color {
+  BLACK = 'black',
+  WHITE = 'white'
+}
