+++ src/model/reversiBoard.ts
@@ -1,6 +1,7 @@
 import Cell from './cell';
 import {Coordinates} from './reversi';
-import Player, {Color} from './player';
+import Player from './player';
+import {Color} from './color.enum';
 
 export default class ReversiBoard {
   private readonly FIELD_SIZE = 8
@@ -28,8 +29,9 @@ export default class ReversiBoard {
     return this.field.some((row) => row.some(cell => cell.isEmpty));
   }
 
-  public markCell(x: number, y: number, player: Player, isFirstPlayer: boolean): void {
+  public markCell(x: number, y: number, player: Player): void {
     const cell = this.field[x][y];
+    const isFirstPlayer = player.color === Color.BLACK;
 
     if (isFirstPlayer) {
       this.firstPlayerCells.push(cell);
@@ -105,7 +107,7 @@ export default class ReversiBoard {
     return this.field.some((row) => row.some(cell => cell.isAvailable));
   }
 
-  public markEarnedEnemyCells(x: number, y: number, currentPlayer: Player, isFirstPlayer: boolean): void {
+  public markEarnedEnemyCells(x: number, y: number, currentPlayer: Player): void {
     const neighborEnemyCellCoords = this.getNeighborEnemyCells(x, y, currentPlayer);
     neighborEnemyCellCoords.forEach((neighborCell) =>{
       const xDiff = neighborCell.x - x;
@@ -118,7 +120,7 @@ export default class ReversiBoard {
         if (this.field[currX][currY].isEmpty) {
           break;
         } else if (this.field[currX][currY].getValue() === currentPlayer) {
-          this.markLineByPlayer({ x, y }, { x: currX, y: currY }, currentPlayer, isFirstPlayer);
+          this.markLineByPlayer({ x, y }, { x: currX, y: currY }, currentPlayer);
           break;
         }
 
@@ -128,7 +130,7 @@ export default class ReversiBoard {
     });
   }
 
-  public markLineByPlayer(start: Coordinates, end: Coordinates, player: Player, isFirstPlayer: boolean): void {
+  public markLineByPlayer(start: Coordinates, end: Coordinates, player: Player): void {
     const xDiff = end.x - start.x > 0 ? 1 : end.x - start.x < 0 ? -1 : 0;
     const yDiff = end.y - start.y > 0 ? 1 : end.y - start.y < 0 ? -1 : 0;
 
@@ -139,7 +141,7 @@ export default class ReversiBoard {
       currY += yDiff;
       currX += xDiff;
 
-      this.markCell(currX, currY, player, isFirstPlayer);
+      this.markCell(currX, currY, player);
     }
   }
