+++ src/players/mcts/MonteCarloAIPlayer.ts
@@ -1,29 +1,31 @@
-import Player, {Color} from '../../model/player';
+import Player from '../../model/player';
 import { Coordinates } from '../../model/reversi';
 import TreeNode from './treeNode';
 import AntiReversiSimulator from './AntiReversiSimulator';
 import ReversiBoard from '../../model/reversiBoard';
+import {Color} from '../../model/color.enum';
 
-export default class SmartAIMonteCarloPlayer extends Player {
-  private MAX_DEPTH: number;
-  private enemy: Player;
+export default class MonteCarloAIPlayer extends Player {
   private currentNode: TreeNode;
   private currentGameSimulation: AntiReversiSimulator;
 
-  constructor(maxDepth: number, color: Color, enemy: Player) {
+  constructor(
+    private maxTime: number,
+    private enemy: Player,
+    color: Color,
+  ) {
     super('SmartAIMonteCarloPlayer', color);
-    this.MAX_DEPTH = maxDepth;
-    this.enemy = enemy;
   }
 
-  getNextMove(inputBoard?: ReversiBoard, maxTime = 1000): Coordinates | Promise<Coordinates> {
-    const antiReversiSimulator = new AntiReversiSimulator(inputBoard, this, this.enemy);
+  getNextMove(inputBoard?: ReversiBoard): Coordinates | Promise<Coordinates> {
+    const antiReversiSimulator = new AntiReversiSimulator(new ReversiBoard(inputBoard), this, this.enemy);
     const root = new TreeNode(null, null, antiReversiSimulator);
-    const startTime = (new Date()).getTime();
-    const timeLimit = startTime + maxTime;
-    for(let iterations = 0; iterations < this.MAX_DEPTH && (new Date()).getTime() < timeLimit; iterations += 1) {
+    const startTime = Date.now();
+    const deadline = startTime + this.maxTime;
+
+    while (Date.now() < deadline) {
       this.currentNode = root;
-      this.currentGameSimulation = new AntiReversiSimulator(inputBoard, this, this.enemy);
+      this.currentGameSimulation = new AntiReversiSimulator(new ReversiBoard(inputBoard), this, this.enemy);
       this.selection();
       this.expansion();
       this.simulation();
@@ -49,10 +51,9 @@ export default class SmartAIMonteCarloPlayer extends Player {
 
   private expansion() {
     if (this.currentNode.uncheckedMoves.length > 0) {
-      //RENAME
-      const actionIndex = Math.floor(Math.random() * this.currentNode.uncheckedMoves.length);
-      this.currentGameSimulation.makeMove(this.currentNode.uncheckedMoves[actionIndex]);
-      this.currentNode = this.currentNode.appendNode(this.currentGameSimulation, actionIndex);
+      const moveIndex = Math.floor(Math.random() * this.currentNode.uncheckedMoves.length);
+      this.currentGameSimulation.makeMove(this.currentNode.uncheckedMoves[moveIndex]);
+      this.currentNode = this.currentNode.appendNode(this.currentGameSimulation, moveIndex);
     }
   }
